+++ openhgnn/dataset/__init__.py
@@ -136,7 +136,8 @@ def build_dataset(dataset, task, *args, **kwargs):
         _dataset = 'rhine_'+task
         return DATASET_REGISTRY[_dataset](dataset, logger=kwargs['logger'],args = kwargs['args'])
     elif dataset in ['dblp4MHGCN','imdb4MHGCN','alibaba4MHGCN']:
-        _dataset = 'mhgcn_' + task    
+        _dataset = 'mhgcn_' + task   
+        return DATASET_REGISTRY[_dataset](dataset, logger=kwargs['logger'],args = kwargs['args']) 
 ##########
 
     elif dataset in ['acm4NSHE', 'acm4GTN', 'academic4HetGNN', 'acm_han', 'acm_han_raw', 'acm4HeCo', 'dblp',
